Simplified Take1 and TakeN into TakeResult
